Cleaner script

+++ doit.sh
@@ -8,13 +8,15 @@ while true; do
     z80asm otrona_d.asm -o otrona_d.bin
     echo "done"
 
-    if ! diff -q otrona_d.bin Otrona_Attaché_U252_Rev_D.BIN; then
-        echo "ERROR!"
-        xxd Otrona_Attaché_U252_Rev_D.BIN > /tmp/a.xxd
-        xxd otrona_d.bin > /tmp/b.xxd
-        diff /tmp/a.xxd /tmp/b.xxd | head -10
-        diff /tmp/b.xxd /tmp/a.xxd | head -10
-    else
-        cp otrona_d.asm otrona_d_lastgood.asm
+    if [ -f otrona_d.bin ]; then
+        if ! diff -q otrona_d.bin Otrona_Attaché_U252_Rev_D.BIN; then
+            echo "ERROR!"
+            xxd Otrona_Attaché_U252_Rev_D.BIN > /tmp/a.xxd
+            xxd otrona_d.bin > /tmp/b.xxd
+            diff /tmp/a.xxd /tmp/b.xxd | head -10
+            # diff /tmp/b.xxd /tmp/a.xxd | head -10
+        else
+            cp otrona_d.asm otrona_d_lastgood.asm
+        fi
     fi
 done
